+++ api/src/main/java/sn/ucad/office/pjobac/modules/security/user/dto/UserRequest.java
@@ -33,7 +33,7 @@ public class UserRequest {
     private  String banque;
     private  String codeBanque;
     private  int codeAgence;
-    private  int numeroCompte;
+    private  Long numeroCompte;
     private  int cleRib;
 
 }
