+++ api/src/main/java/sn/ucad/office/pjobac/modules/security/user/dto/UserResponse.java
@@ -34,7 +34,7 @@ public class UserResponse {
     private  String banque;
     private  String codeBanque;
     private  int codeAgence;
-    private  int numeroCompte;
+    private  Long numeroCompte;
     private  int cleRib;
     private boolean isEnabled;
     private boolean isLocked;
